feat: add registration page for company with form fields and navigation feat: replace favicon.ico with favicon.svg feat: update logo.svg with new design feat: add star.svg icon for UI components
